git init defaults to "main", added support for local branch
As it turns out, "master" is hardcoded into git. So, getting around that
involves intercepting calls to `git`, checking to see if they use
`init`, and then configuring "main" instead.

Big thanks to this blog post for that one:
https://andre.arko.net/2020/06/06/changing-git-and-githubs-default-branch-name/

Also, ~/.zsh-local.zsh is now read in and executed to extend zshrc. This
is so that I can have machine-specific configs.